Safety Commit
Just in case I need to revert back to this later.
