+++ Assets/New Moveset.asset
@@ -10,13 +10,10 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 2ed0e1784cd049cb979931efd12ea662, type: 3}
-  m_Name: Reaper Moveset
+  m_Name: New Moveset
   m_EditorClassIdentifier: 
   moves: []
-  punchMoves:
-  - {fileID: 11400000, guid: e1becc4c3ea6d34479f7194b077262b6, type: 2}
-  kickMoves:
-  - {fileID: 11400000, guid: cb3ab6d572c724440b50af7950a10e51, type: 2}
-  slashMoves:
-  - {fileID: 11400000, guid: 5bd417a604b16544c8ab809093c5eeb0, type: 2}
+  punchMoves: []
+  kickMoves: []
+  slashMoves: []
   uniqueMoves: []
