+++ Assets/New Moveset.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 48ddffbf4014a224c84cfebb6b5bd28f
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
