+++ Assets/_Project/Editor/Scripts/MovesetCreator.cs.meta
@@ -1,3 +1,3 @@
 ﻿fileFormatVersion: 2
-guid: 7e2df70f394a4cc99dd13778ae929f60
-timeCreated: 1696124571
+guid: 724513873ad74de98a59a0e1d06369d4
+timeCreated: 1696160518
